remove app directory
